+++ lib/model/network_device.dart
@@ -1,15 +1,9 @@
-
-
 class NetworkDevice {
   final String? name;
   final String address;
   final int port;
 
-  NetworkDevice({
-    this.name,
-    required this.address,
-    required this.port,
-  });
+  NetworkDevice({this.name, required this.address, required this.port});
 
   @override
   String toString() {