+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 1.0.0
 
 * Initial release of the project.
+
+## 1.0.1
+
+* Static analysis update.
+
+## 1.0.2
+
+* Enhance stream to use stream controller.